:recycle: refactor: rename `nextauthUser` to `nextAuthUser`
